Disable CosmosPlugin for AF3
